t1disasm: Fix infinite loop present in some fonts.
Niels Thykier reported a problem with a Chinese-language font
bundled with Debian: t1disasm entered an infinite loop printing
"2 index /CharStrings...". The problem was our line extension
protocol; we missed a case where we should've reset the saved
line to empty. Fix this, and make it less likely in future by
simplifying the saved-line protocol.